+++ returnn/config.py
@@ -14,29 +14,10 @@ from sisyphus import *
 from sisyphus.delayed_ops import DelayedBase
 from sisyphus.hash import sis_hash_helper
 
+from i6_core.util import instanciate_delayed
+
 Path = setup_path(__package__)
 Variable = tk.Variable
-
-
-def instanciate_delayed(o):
-    """
-    Recursively traverses a structure and calls .get() on all
-    existing Delayed Operations, especially Variables in the structure
-
-    :param Any o: nested structure that may contain DelayedBase objects
-    :return:
-    """
-    if isinstance(o, DelayedBase):
-        o = o.get()
-    elif isinstance(o, list):
-        for k in range(len(o)):
-            o[k] = instanciate_delayed(o[k])
-    elif isinstance(o, tuple):
-        o = tuple(instanciate_delayed(e) for e in o)
-    elif isinstance(o, dict):
-        for k in o:
-            o[k] = instanciate_delayed(o[k])
-    return o
 
 
 class CodeWrapper:
